==> python scripts/2_rc2015updateDEMdata.ipynb <==
# Import all libraries needed for the tutorial
import pandas as pd
import pymysql
from pandas.io import sql
from pandas.io.sql import read_sql, to_sql
from sqlalchemy import create_engine, Table, Column, Integer, String, MetaData, ForeignKey

#pd.set_option('display.mpl_style', 'default') 
pd.set_option('display.width', 5000) 
pd.set_option('display.max_columns', 60)
pd.set_option('display.max_rows', 60)

print('is loaded')

# loads in the data in manageable chunks, making sure there's no header 

base = pd.read_csv('rc15csv/rc15dem.txt', sep=';', header=None, iterator=True, chunksize=1000, low_memory=False)
data = pd.concat(list(base), ignore_index=True)

print('this is loaded')

data.shape

data.head(3)

# update Recurring
recurr2015 = data[[0,12,13,20,36,53,253,255,256]]
recurr2015.columns = ['schid','grades2015','demschoolwhite','enroll2015','enrolldist2015','lowinc2015','act2015school','actscore2015district','actscore2015state']
recurr2015 = recurr2015.applymap(lambda x: x.lstrip(' ').rstrip(' ') if isinstance(x, str) else x)
recurr2015['percmin2015'] = (100 - recurr2015['demschoolwhite'])
recurr2015.percmin2015 = recurr2015.percmin2015.round(1)
recurr2015 = recurr2015[['schid','grades2015','percmin2015','enroll2015','enrolldist2015','lowinc2015','act2015school','actscore2015district','actscore2015state']]
recurr2015.head(3)

recurr2015.to_csv("rc15csv/UPDATErecurr2015.csv", index=False)

con = create_engine('mysql+pymysql://root:root@localhost:8889/School Report Cards', echo=False)
print('connected')

#where values are all spaces inserts --
data = data.applymap(lambda x: "--" if isinstance(x, str) and x.isspace() else x)
# strip leading and trailing spaces
data = data.applymap(lambda x: x.lstrip(' ').rstrip(' ') if isinstance(x, str) else x)
# Where values are empty NaN, fills them with --
data = data.fillna('--')
data.head(3)

rc2015dem_data = data[[0,13,14,15,16,17,18,19,29,30,31,32,33,34,35,37,38,39,40,41,42,43,44,45,47,48,49,51,52,55,56,57,59,60,65,67,68,69,71,72,125,127,128,133,135,136,137,139,140,141,143,144,145,147,148,149,151,152,153,155,156,157,159,160,161,163,164,165,167,168,169,171,172,173,175,176,177,179,180,181,183,184,185,187,188,189,191,192,193,195,196,197,199,200,201,203,204,205,207,208,209,211,212,213,215,216,217,219,220,221,223,224,225,227,228,229,231,232,233,235,236,237,239,240,241,243,244,245,247,248,249,251,252,257,259,260,261,263,264,265,267,268,269,271,272,273,275,276,277,279,280,281,283,284,285,287,288,289,291,292,293,295,296,309,311,312,313,315,316,317,319,320,321,323,324,325,327,328,329,331,332,333,335,336,389,390,391,392,393,394,395,396,397,398,399,400,401,402,403,404,405,406,407,408,409,410,411,412,419,420,427,428,435,436,439,440,451,452,455,456,475,477,478,479,481,482,484,485,487,488,490,491,493,494,496,497,505,506,509,510,513,514,515,516,517,518,519,520,521,522,523,524,525,526,527,528,529,530,531,532,533,534,535,536,537,538,539,540,541,542,543,544,545,546,547,548,737,739,740,741,743,744,745,747,748,749,751,752]]
rc2015dem_data.columns = ['schid','demschoolwhite','demschoolblack','demschoolhisp','demschoolasian','demschoolnathaw','demschoolnatam','demschoolmulti','demdistwhite','demdistblack','demdisthisp','demdistasian','demdistnathaw','demdistnatam','demdistmulti','demstatewhite','demstateblack','demstatehisp','demstateasian','demstatenathaw','demstatenatam','demstatemulti','enrollstate','demschoollep','demdistlep','demstatelep','demschooliep','demdistiep','demstateiep','demdistlow','demstatelow','demschoolhome','demdisthome','demstatehome','demschoolparent','demdistparent','demstateparent','percattendschoolall','percattenddistall','percattendstateall','percmobileschool','percmobiledist','percmobilestate','perctruantschool','perctruantdist','perctruantstate','percdropoutschool','percdropoutdist','percdropoutstate','perchs4gradschoolall','perchs4graddistall','perchs4gradstateall','perchs4gradschoolmale','perchs4graddistmale','perchs4gradstatemale','perchs4gradschoolfemale','perchs4graddistfemale','perchs4gradstatefemale','perchs4gradschoolwhite','perchs4graddistwhite','perchs4gradstatewhite','perchs4gradschoolblack','perchs4graddistblack','perchs4gradstateblack','perchs4gradschoolhisp','perchs4graddisthisp','perchs4gradstatehisp','perchs4gradschoolasian','perchs4graddistasian','perchs4gradstateasian','perchs4gradschoolnathaw','perchs4graddistnathaw','perchs4gradstatenathaw','perchs4gradschoolnatam','perchs4graddistnatam','perchs4gradstatenatam','perchs4gradschoolmulti','perchs4graddistmulti','perchs4gradstatemulti','perchs4gradschoollep','perchs4graddistlep','perchs4gradstatelep','perchs4gradschoolmigrant','perchs4graddistmigrant','perchs4gradstatemigrant','perchs4gradschooliep','perchs4graddistiep','perchs4gradstateiep','perchs4gradschoollow','perchs4graddistlow','perchs4gradstatelow','perchs5gradschoolall','perchs5graddistall','perchs5gradstateall','perchs5gradschoolmale','perchs5graddistmale','perchs5gradstatemale','perchs5gradschoolfemale','perchs5graddistfemale','perchs5gradstatefemale','perchs5gradschoolwhite','perchs5graddistwhite','perchs5gradstatewhite','perchs5gradschoolblack','perchs5graddistblack','perchs5gradstateblack','perchs5gradschoolhisp','perchs5graddisthisp','perchs5gradstatehisp','perchs5gradschoolasian','perchs5graddistasian','perchs5gradstateasian','perchs5gradschoolnathaw','perchs5graddistnathaw','perchs5gradstatenathaw','perchs5gradschoolnatam','perchs5graddistnatam','perchs5gradstatenatam','perchs5gradschoolmulti','perchs5graddistmulti','perchs5gradstatemulti','perchs5gradschoollep','perchs5graddistlep','perchs5gradstatelep','perchs5gradschoolmigrant','perchs5graddistmigrant','perchs5gradstatemigrant','perchs5gradschooliep','perchs5graddistiep','perchs5gradstateiep','perchs5gradschoollow','perchs5graddistlow','perchs5gradstatelow','actschooleng','actdisteng','actstateeng','actschoolmath','actdistmath','actstatemath','actschoolread','actdistread','actstateread','actschoolsci','actdistsci','actstatesci','percschoolcol','percdistcol','percstatecol','actschoolengmet','actdistengmet','actstateengmet','actschoolmathmet','actdistmathmet','actstatemathmet','actschoolreadmet','actdistreadmet','actstatereadmet','actschoolscimet','actdistscimet','actstatescimet','actschoolallmet','actdistallmet','actstateallmet','sizeschool3','sizedist3','sizestate3','sizeschool4','sizedist4','sizestate4','sizeschool5','sizedist5','sizestate5','sizeschool6','sizedist6','sizestate6','sizeschool7','sizedist7','sizestate7','sizeschool8','sizedist8','sizestate8','sizeschoolhs','sizedisths','sizestatehs','teacherdistpercwhite','teacherstatepercwhite','teacherdistpercblack','teacherstatepercblack','teacherdistperchisp','teacherstateperchisp','teacherdistpercasian','teacherstatepercasian','teacherdistpercnathaw','teacherstatepercnathaw','teacherdistpercnatam','teacherstatepercnatam','teacherdistpercmulti','teacherstatepercmulti','teacherdistpercunknown','teacherstatepercunknown','teacherdistpercmale','teacherstatepercmale','teacherdistpercfemale','teacherstatepercfemale','teacherdisttotalfte','teacherstatetotalfte','teacherdistavgexp','teacherstateavgexp','teacherdistbach','teacherstatebach','teacherdistmast','teacherstatemast','teacherratiodistelem','teacherratiostateelem','teacherratiodisths','teacherratiostatehs','teachsaldist','teachsalstate','adminsaldist','adminsalstate','teacherschoolret','teacherdistret','teacherstateret','principalschoolturn','principaldistturn','principalstateturn','fundlocpropdistperc','fundlocpropstateperc','fundotherdistperc','fundotherstateperc','fundstateaiddistperc','fundstateaidstateperc','fundotherstatedistperc','fundotherstatestateperc','fundfeddistperc','fundfedstateperc','instexppupildist','instexppupilstate','otherexppupildist','otherexppupilstate','instdollars','instdollarsdistperc','instdollarsstateperc','genadmindollars','genadmindistperc','genadminstateperc','supportservdollars','supportservdistperc','supportservstateperc','otherexpdollars','otherexpdistperc','otherexpstateperc','edfunddollars','edfunddistperc','edfundstateperc','opbmdollars','opbmdistperc','opbmstateperc','transportdollars','transportdistperc','transportstateperc','debtservdollars','debtservdistperc','debtservstateperc','tortdollars','tortdistperc','tortstateperc','muncretsocdollars','muncretsocdistperc','muncretsocstateperc','fireprevdollars','fireprevdistperc','fireprevstateperc','capdollars','capdistperc','capstateperc','percschoolenroll16','percdistenroll16','percstateenroll16','percschoolenroll12','percdistenroll12','percstateenroll12','percschoolfresh','percdistfresh','percstatefresh','perc8algsch','perc8algdist','perc8algstate']
rc2015dem_data.head(3)

nottested2015 = pd.read_csv('rc15csv/nottested2015.csv', 
                      converters={'schid': str},
                      engine='python', 
                      index_col=None)
nottested2015.head(3)

left = rc2015dem_data
right = nottested2015
update2015 = pd.merge(left, right, how='left', on='schid', suffixes=('_x', '_y'))
update2015.shape

dem_columns=[
    Column('schid', String(20), primary_key=True),
    Column('demschoolwhite', String(5)),
    Column('demschoolblack', String(5)),
    Column('demschoolhisp', String(5)),
    Column('demschoolasian', String(5)),
    Column('demschoolnathaw', String(5)),
    Column('demschoolnatam', String(5)),
    Column('demschoolmulti', String(5)),
    Column('demdistwhite', String(5)),
    Column('demdistblack', String(5)),
    Column('demdisthisp', String(5)),
    Column('demdistasian', String(5)),
    Column('demdistnathaw', String(5)),
    Column('demdistnatam', String(5)),
    Column('demdistmulti', String(5)),
    Column('demstatewhite', String(5)),
    Column('demstateblack', String(5)),
    Column('demstatehisp', String(5)),
    Column('demstateasian', String(5)),
    Column('demstatenathaw', String(5)),
    Column('demstatenatam', String(5)),
    Column('demstatemulti', String(5)),
    Column('enrollstate', String(10)),
    Column('demschoollep', String(5)),
    Column('demdistlep', String(5)),
    Column('demstatelep', String(5)),
    Column('demschooliep', String(5)),
    Column('demdistiep', String(5)),
    Column('demstateiep', String(5)),
    Column('demdistlow', String(5)),
    Column('demstatelow', String(5)),
    Column('demschoolhome', String(5)),
    Column('demdisthome', String(5)),
    Column('demstatehome', String(5)),
    Column('demschoolparent', String(5)),
    Column('demdistparent', String(5)),
    Column('demstateparent', String(5)),
    Column('percattendschoolall', String(5)),
    Column('percattenddistall', String(5)),
    Column('percattendstateall', String(5)),
    Column('percmobileschool', String(5)),
    Column('percmobiledist', String(5)),
    Column('percmobilestate', String(5)),
    Column('perctruantschool', String(5)),
    Column('perctruantdist', String(5)),
    Column('perctruantstate', String(5)),
    Column('percdropoutschool', String(5)),
    Column('percdropoutdist', String(5)),
    Column('percdropoutstate', String(5)),
    Column('perchs4gradschoolall', String(5)),
    Column('perchs4graddistall', String(5)),
    Column('perchs4gradstateall', String(5)),
    Column('perchs4gradschoolmale', String(5)),
    Column('perchs4graddistmale', String(5)),
    Column('perchs4gradstatemale', String(5)),
    Column('perchs4gradschoolfemale', String(5)),
    Column('perchs4graddistfemale', String(5)),
    Column('perchs4gradstatefemale', String(5)),
    Column('perchs4gradschoolwhite', String(5)),
    Column('perchs4graddistwhite', String(5)),
    Column('perchs4gradstatewhite', String(5)),
    Column('perchs4gradschoolblack', String(5)),
    Column('perchs4graddistblack', String(5)),
    Column('perchs4gradstateblack', String(5)),
    Column('perchs4gradschoolhisp', String(5)),
    Column('perchs4graddisthisp', String(5)),
    Column('perchs4gradstatehisp', String(5)),
    Column('perchs4gradschoolasian', String(5)),
    Column('perchs4graddistasian', String(5)),
    Column('perchs4gradstateasian', String(5)),
    Column('perchs4gradschoolnathaw', String(5)),
    Column('perchs4graddistnathaw', String(5)),
    Column('perchs4gradstatenathaw', String(5)),
    Column('perchs4gradschoolnatam', String(5)),
    Column('perchs4graddistnatam', String(5)),
    Column('perchs4gradstatenatam', String(5)),
    Column('perchs4gradschoolmulti', String(5)),
    Column('perchs4graddistmulti', String(5)),
    Column('perchs4gradstatemulti', String(5)),
    Column('perchs4gradschoollep', String(5)),
    Column('perchs4graddistlep', String(5)),
    Column('perchs4gradstatelep', String(5)),
    Column('perchs4gradschoolmigrant', String(5)),
    Column('perchs4graddistmigrant', String(5)),
    Column('perchs4gradstatemigrant', String(5)),
    Column('perchs4gradschooliep', String(5)),
    Column('perchs4graddistiep', String(5)),
    Column('perchs4gradstateiep', String(5)),
    Column('perchs4gradschoollow', String(5)),
    Column('perchs4graddistlow', String(5)),
    Column('perchs4gradstatelow', String(5)),
    Column('perchs5gradschoolall', String(5)),
    Column('perchs5graddistall', String(5)),
    Column('perchs5gradstateall', String(5)),
    Column('perchs5gradschoolmale', String(5)),
    Column('perchs5graddistmale', String(5)),
    Column('perchs5gradstatemale', String(5)),
    Column('perchs5gradschoolfemale', String(5)),
    Column('perchs5graddistfemale', String(5)),
    Column('perchs5gradstatefemale', String(5)),
    Column('perchs5gradschoolwhite', String(5)),
    Column('perchs5graddistwhite', String(5)),
    Column('perchs5gradstatewhite', String(5)),
    Column('perchs5gradschoolblack', String(5)),
    Column('perchs5graddistblack', String(5)),
    Column('perchs5gradstateblack', String(5)),
    Column('perchs5gradschoolhisp', String(5)),
    Column('perchs5graddisthisp', String(5)),
    Column('perchs5gradstatehisp', String(5)),
    Column('perchs5gradschoolasian', String(5)),
    Column('perchs5graddistasian', String(5)),
    Column('perchs5gradstateasian', String(5)),
    Column('perchs5gradschoolnathaw', String(5)),
    Column('perchs5graddistnathaw', String(5)),
    Column('perchs5gradstatenathaw', String(5)),
    Column('perchs5gradschoolnatam', String(5)),
    Column('perchs5graddistnatam', String(5)),
    Column('perchs5gradstatenatam', String(5)),
    Column('perchs5gradschoolmulti', String(5)),
    Column('perchs5graddistmulti', String(5)),
    Column('perchs5gradstatemulti', String(5)),
    Column('perchs5gradschoollep', String(5)),
    Column('perchs5graddistlep', String(5)),
    Column('perchs5gradstatelep', String(5)),
    Column('perchs5gradschoolmigrant', String(5)),
    Column('perchs5graddistmigrant', String(5)),
    Column('perchs5gradstatemigrant', String(5)),
    Column('perchs5gradschooliep', String(5)),
    Column('perchs5graddistiep', String(5)),
    Column('perchs5gradstateiep', String(5)),
    Column('perchs5gradschoollow', String(5)),
    Column('perchs5graddistlow', String(5)),
    Column('perchs5gradstatelow', String(5)),
    Column('actschooleng', String(4)),
    Column('actdisteng', String(4)),
    Column('actstateeng', String(4)),
    Column('actschoolmath', String(4)),
    Column('actdistmath', String(4)),
    Column('actstatemath', String(4)),
    Column('actschoolread', String(4)),
    Column('actdistread', String(4)),
    Column('actstateread', String(4)),
    Column('actschoolsci', String(4)),
    Column('actdistsci', String(4)),
    Column('actstatesci', String(4)),
    Column('percschoolcol', String(5)),
    Column('percdistcol', String(5)),
    Column('percstatecol', String(5)),
    Column('actschoolengmet', String(5)),
    Column('actdistengmet', String(5)),
    Column('actstateengmet', String(5)),
    Column('actschoolmathmet', String(5)),
    Column('actdistmathmet', String(5)),
    Column('actstatemathmet', String(5)),
    Column('actschoolreadmet', String(5)),
    Column('actdistreadmet', String(5)),
    Column('actstatereadmet', String(5)),
    Column('actschoolscimet', String(5)),
    Column('actdistscimet', String(5)),
    Column('actstatescimet', String(5)),
    Column('actschoolallmet', String(5)),
    Column('actdistallmet', String(5)),
    Column('actstateallmet', String(5)),
    Column('sizeschool3', String(5)),
    Column('sizedist3', String(5)),
    Column('sizestate3', String(5)),
    Column('sizeschool4', String(5)),
    Column('sizedist4', String(5)),
    Column('sizestate4', String(5)),
    Column('sizeschool5', String(5)),
    Column('sizedist5', String(5)),
    Column('sizestate5', String(5)),
    Column('sizeschool6', String(5)),
    Column('sizedist6', String(5)),
    Column('sizestate6', String(5)),
    Column('sizeschool7', String(5)),
    Column('sizedist7', String(5)),
    Column('sizestate7', String(5)),
    Column('sizeschool8', String(5)),
    Column('sizedist8', String(5)),
    Column('sizestate8', String(5)),
    Column('sizeschoolhs', String(5)),
    Column('sizedisths', String(5)),
    Column('sizestatehs', String(5)),
    Column('teacherdistpercwhite', String(5)),
    Column('teacherstatepercwhite', String(5)),
    Column('teacherdistpercblack', String(5)),
    Column('teacherstatepercblack', String(5)),
    Column('teacherdistperchisp', String(5)),
    Column('teacherstateperchisp', String(5)),
    Column('teacherdistpercasian', String(5)),
    Column('teacherstatepercasian', String(5)),
    Column('teacherdistpercnathaw', String(5)),
    Column('teacherstatepercnathaw', String(5)),
    Column('teacherdistpercnatam', String(5)),
    Column('teacherstatepercnatam', String(5)),
    Column('teacherdistpercmulti', String(5)),
    Column('teacherstatepercmulti', String(5)),
    Column('teacherdistpercunknown', String(5)),
    Column('teacherstatepercunknown', String(5)),
    Column('teacherdistpercmale', String(5)),
    Column('teacherstatepercmale', String(5)),
    Column('teacherdistpercfemale', String(5)),
    Column('teacherstatepercfemale', String(5)),
    Column('teacherdisttotalfte', String(7)),
    Column('teacherstatetotalfte', String(7)),
    Column('teacherdistavgexp', String(5)),
    Column('teacherstateavgexp', String(5)),
    Column('teacherdistbach', String(5)),
    Column('teacherstatebach', String(5)),
    Column('teacherdistmast', String(5)),
    Column('teacherstatemast', String(5)),
    Column('teacherratiodistelem', String(5)),
    Column('teacherratiostateelem', String(5)),
    Column('teacherratiodisths', String(5)),
    Column('teacherratiostatehs', String(5)),
    Column('teachsaldist', String(11)),
    Column('teachsalstate', String(11)),
    Column('adminsaldist', String(11)),
    Column('adminsalstate', String(11)),
    Column('teacherschoolret', String(5)),
    Column('teacherdistret', String(5)),
    Column('teacherstateret', String(5)),
    Column('principalschoolturn', String(4)),
    Column('principaldistturn', String(4)),
    Column('principalstateturn', String(4)),
    Column('fundlocpropdistperc', String(5)),
    Column('fundlocpropstateperc', String(5)),
    Column('fundotherdistperc', String(5)),
    Column('fundotherstateperc', String(5)),
    Column('fundstateaiddistperc', String(5)),
    Column('fundstateaidstateperc', String(5)),
    Column('fundotherstatedistperc', String(5)),
    Column('fundotherstatestateperc', String(5)),
    Column('fundfeddistperc', String(5)),
    Column('fundfedstateperc', String(5)),
    Column('instexppupildist', String(10)),
    Column('instexppupilstate', String(10)),
    Column('otherexppupildist', String(10)),
    Column('otherexppupilstate', String(10)),
    Column('instdollars', String(16)),
    Column('instdollarsdistperc', String(5)),
    Column('instdollarsstateperc', String(5)),
    Column('genadmindollars', String(16)),
    Column('genadmindistperc', String(5)),
    Column('genadminstateperc', String(5)),
    Column('supportservdollars', String(16)),
    Column('supportservdistperc', String(5)),
    Column('supportservstateperc', String(5)),
    Column('otherexpdollars', String(16)),
    Column('otherexpdistperc', String(5)),
    Column('otherexpstateperc', String(5)),
    Column('edfunddollars', String(16)),
    Column('edfunddistperc', String(5)),
    Column('edfundstateperc', String(5)),
    Column('opbmdollars', String(16)),
    Column('opbmdistperc', String(5)),
    Column('opbmstateperc', String(5)),
    Column('transportdollars', String(16)),
    Column('transportdistperc', String(5)),
    Column('transportstateperc', String(5)),
    Column('debtservdollars', String(16)),
    Column('debtservdistperc', String(5)),
    Column('debtservstateperc', String(5)),
    Column('tortdollars', String(16)),
    Column('tortdistperc', String(5)),
    Column('tortstateperc', String(5)),
    Column('muncretsocdollars', String(16)),
    Column('muncretsocdistperc', String(5)),
    Column('muncretsocstateperc', String(5)),
    Column('fireprevdollars', String(16)),
    Column('fireprevdistperc', String(5)),
    Column('fireprevstateperc', String(5)),
    Column('capdollars', String(16)),
    Column('capdistperc', String(5)),
    Column('capstateperc', String(5)),
    Column('percschoolenroll16', String(5)),
    Column('percdistenroll16', String(5)),
    Column('percstateenroll16', String(5)),
    Column('percschoolenroll12', String(5)),
    Column('percdistenroll12', String(5)),
    Column('percstateenroll12', String(5)),
    Column('percschoolfresh', String(5)),
    Column('percdistfresh', String(5)),
    Column('percstatefresh', String(5)),
    Column('perc8algsch', String(5)),
    Column('perc8algdist', String(5)),
    Column('perc8algstate', String(5)),
    Column('nttkreadsch', String(5)),
    Column('nttkreaddist', String(5)),
    Column('nttkreadstate', String(5)),
    Column('nttkmathsch', String(5)),
    Column('nttkmathdist', String(5)),
    Column('nttkmathstate', String(5)),
    Column('tkreadsch', String(5)),
    Column('tkmathsch', String(5)),
    Column('tkreaddist', String(5)),
    Column('tkmathdist', String(5)),
    Column('tkreadstate', String(5)),
    Column('tkmathstate', String(5))
      ]
print('ready')

metadata = MetaData()
rc2015dem = Table('rc2015dem', metadata, *dem_columns, mysql_engine='InnoDB')

print('metadata created')

metadata.create_all(con)
print("created")

#drops tables if they exist
con.execute("DROP TABLE IF EXISTS rc2015dem")# for demographics

#uploads all the data
update2015.to_sql(name='rc2015dem', con=con, if_exists = 'append', index=False, index_label=None, chunksize=1000)
print('done')

update2015.to_csv("rc15csv/UPDATErc2015dem.csv", index=False)
