==> python scripts/3_JoinUPDATErecurr2015scores.ipynb <==
import pandas as pd
import numpy as np
print("ready")

sch2015only = pd.read_csv('rc15csv/rc2015sch.csv', index_col=None)
sch2015only.head(3)

sch2015only = sch2015only.drop('Unnamed: 214',1)
sch2015only = sch2015only.drop('Unnamed: 215',1)
sch2015only.shape

updaterecurr = pd.read_csv('rc15csv/UPDATErecurr2015.csv', index_col=None)
print('1 loaded')

updaterecurr.head(3)

left = sch2015only
right = updaterecurr
result = pd.merge(left, right, how='left', on='schid', suffixes=('_x', '_y'))
result.shape

result.head(3)

me2015 = pd.read_csv('rc15csv/me2015.csv', index_col=None)
me2015.head()

left = result
right = me2015
sch2015me = pd.merge(left, right, how='left', on='schid', suffixes=('_x', '_y'))
sch2015me.shape

#'ELEMENTARY','MIDDLE SCHL','CHARTER SCH','CHARTER NET','CHARTER NET SCH','HIGH SCHOOL','CHARTER HIGH SCH','CHARTER NET HIGH SCH','CHARTER HIGH SCH+','CHARTER NET HIGH SCH+','HIGH SCHOOL+','OTHER','PRESCHOOL','EARLY CHILDHOOD'

#Selects only rows for elmentary schools, with charter sch included in elementary
elemval = ['HIGH SCHOOL','CHARTER HIGH SCH','CHARTER NET','CHARTER NET+','CHARTER NET HIGH SCH','CHARTER HIGH SCH+','CHARTER NET HIGH SCH+','HIGH SCHOOL+','OTHER','PRESCHOOL','EARLY CHILDHOOD']
me2015elem = sch2015me[~sch2015me.schooltype.isin(elemval)]

#renames column for elementary school scores
me2015elem = me2015elem.rename(columns = {'me2015school':'me2015schoolELEM'})
me2015elem = me2015elem[['schid','me2015schoolELEM']]

#Selects only rows for high schools
hsval = ['ELEMENTARY','MIDDLE SCHL','CHARTER SCH','CHARTER SCH+','CHARTER NET','CHARTER NET+','CHARTER NET SCH','OTHER','PRESCHOOL','EARLY CHILDHOOD']
me2015hs = sch2015me[~sch2015me.schooltype.isin(hsval)]

#renames column for elementary school scores
me2015hs = me2015hs.rename(columns = {'me2015school':'me2015schoolHS'})
me2015hs = me2015hs[['schid','me2015schoolHS']]

left = sch2015me
right = me2015elem
resultElem = pd.merge(left, right, how='left', on='schid', suffixes=('_x', '_y'))
resultElem.shape

left = resultElem
right = me2015hs
resultall = pd.merge(left, right, how='left', on='schid', suffixes=('_x', '_y'))
resultall.shape

resultall = resultall.applymap(lambda x: "" if isinstance(x, str) and x.isspace() else x)
resultall = resultall.fillna('')
resultall.head(3)

resultall.shape

resultall.to_csv("rc15csv/pai2015before.csv", index=False)
