==> python scripts/4_figurePAIndexCOMBINE.ipynb <==
import pandas as pd
import numpy as np

pd.options.mode.chained_assignment = None  # default='warn'

print('This is loaded')

base = pd.read_csv('rc15csv/pai2015before.csv', thousands=",", index_col='schid')
print('1 loaded')

base.head(3)

base['zelemscore2015me'] = 500.0
base['avgelem2015me'] = ''
base['stdvelem2015me'] = ''
base['zelemscore2015meapp'] = 500.0
base['avgelem2015meapp'] = ''
base['stdvelem2015meapp'] = ''
base['avgELEMz'] = 500.0
base['zhsscore2015me'] = 500.0
base['avghs2015me'] = ''
base['stdvhs2015me'] = ''
base['zhsscore2015meapp'] = 500.0
base['avghs2015meapp'] = ''
base['stdvhs2015meapp'] = ''
base['avgHSz'] = 500.0

base.shape

elemval = ['CHARTER NET','CHARTER NET+']
elembase = base[~base.schooltype.isin(elemval)]

elembase['MEappelem'] = elembase['me2015schoolELEM'] + elembase['me2015schAPP']
elembase.shape

#figure elem schools

lowincrange = {0: 10, 10: 20, 20: 30, 30: 40, 40: 50, 50: 60, 60: 70, 70: 80, 80: 90, 90: 101}

for key in lowincrange:
    dfselect = elembase[ elembase["me2015schoolELEM"].notnull() & ( elembase['lowinc2015'] >= key ) & ( elembase['lowinc2015'] < lowincrange[key] ) ]

    meavg = round(dfselect["me2015schoolELEM"].mean(),1)
    mestd = round(dfselect["me2015schoolELEM"].std(ddof=1),2)
    dfselect.loc[:, 'avgelem2015me'] = meavg
    dfselect.loc[:, 'stdvelem2015me'] = mestd    
    dfselect.loc[:,'zelemscore2015me'] = ( ( dfselect['me2015schoolELEM'] - meavg) / mestd ).round(2)
    ssize = len(dfselect)
    sterror = round((mestd / np.sqrt(ssize)) * 1.98,2)
    low95 = round(meavg - sterror,2)
    high95 = round(meavg + sterror,2)

    meAPPavg = round(dfselect["MEappelem"].mean(),1)
    meAPPstd = round(dfselect["MEappelem"].std(ddof=1),2)
    dfselect.loc[:, 'avgelem2015meapp'] = meavg
    dfselect.loc[:, 'stdvelem2015meapp'] = mestd    
    dfselect.loc[:,'zelemscore2015meapp'] = ( ( dfselect['MEappelem'] - meAPPavg) / meAPPstd ).round(2)
    sterrorAPP = round((meAPPstd / np.sqrt(ssize)) * 1.98,2)
    low95APP = round(meAPPavg - sterror,2)
    high95APP = round(meAPPavg + sterror,2)

    dfselect.loc[:,'avgELEMz'] = ( ( dfselect['zelemscore2015meapp'] + dfselect['zelemscore2015me'] ) /2).round(2)

    print("Range ME | ME | Sample | StDev | Error | 95% Low | 95% High")
    print( key,"to",lowincrange[key]," | ",meavg," | ",ssize," | ",mestd," | ",sterror," | ",low95," | ",high95 )
    print("Range ME+APP | ME+APP | Sample | StDev | Error | 95% Low | 95% High")
    print( key,"to",lowincrange[key]," | ",meAPPavg," | ",ssize," | ",meAPPstd," | ",sterrorAPP," | ",low95APP," | ",high95APP )

    base.update(dfselect)
    
print('done')

base.shape

base['MEappHS'] = base['me2015schoolHS'] + base['me2015schAPP']
base.shape

#figure high schools

lowincrange = {0: 12.5, 12.5: 25, 25: 37.5, 37.5: 50, 50: 62.5, 62.5: 75, 75: 87.5, 87.5: 101}

for key in lowincrange:
    dfselect = base[ base["me2015schoolHS"].notnull() & ( base['lowinc2015'] >= key ) & ( base['lowinc2015'] < lowincrange[key] ) ]

    meavg = round(dfselect["me2015schoolHS"].mean(),1)
    mestd = round(dfselect["me2015schoolHS"].std(ddof=1),2)
    dfselect.loc[:, 'avghs2015me'] = meavg
    dfselect.loc[:, 'stdvhs2015me'] = mestd
    dfselect.loc[:,'zhsscore2015me'] = ( ( dfselect['me2015schoolHS'] - meavg) / mestd ).round(2)
    ssize = len(dfselect)
    sterror = round((mestd / np.sqrt(ssize)) * 1.98,2)
    low95 = round(meavg - sterror,2)
    high95 = round(meavg + sterror,2)

    meAPPavg = round(dfselect["MEappHS"].mean(),1)
    meAPPstd = round(dfselect["MEappHS"].std(ddof=1),2)
    dfselect.loc[:, 'avghs2015meapp'] = meavg
    dfselect.loc[:, 'stdvhs2015meapp'] = mestd    
    dfselect.loc[:,'zhsscore2015meapp'] = ( ( dfselect['MEappHS'] - meAPPavg) / meAPPstd ).round(2)
    sterrorAPP = round((meAPPstd / np.sqrt(ssize)) * 1.98,2)
    low95APP = round(meAPPavg - sterror,2)
    high95APP = round(meAPPavg + sterror,2)

    dfselect.loc[:,'avgHSz'] = ( ( dfselect['zhsscore2015meapp'] + dfselect['zhsscore2015me'] ) /2).round(2)
    
    print("Range ME | ME | Sample | StDev | Error | 95% Low | 95% High")
    print( key,"to",lowincrange[key]," | ",meavg," | ",ssize," | ",mestd," | ",sterror," | ",low95," | ",high95 )
    print("Range ME+APP | ME+APP | Sample | StDev | Error | 95% Low | 95% High")
    print( key,"to",lowincrange[key]," | ",meAPPavg," | ",ssize," | ",meAPPstd," | ",sterrorAPP," | ",low95APP," | ",high95APP )

    base.update(dfselect)
    
print('done')

base.head(3)

base.loc[base['zelemscore2015me'] == 500.0, 'zelemscore2015me'] = ''
base.loc[base['zelemscore2015meapp'] == 500.0, 'zelemscore2015meapp'] = ''
base.loc[base['avgELEMz'] == 500.0, 'avgELEMz'] = ''
base.loc[base['zhsscore2015me'] == 500.0, 'zhsscore2015me'] = ''
base.loc[base['zhsscore2015meapp'] == 500.0, 'zhsscore2015meapp'] = ''
base.loc[base['avgHSz'] == 500.0, 'avgHSz'] = ''
base.head(3)

base = base.drop('MEappHS', 1)
base.shape

#for me to fiddle with
base.to_csv("rc15csv/pai2015done.csv")
