==> python scripts/5_load2015schools.ipynb <==
# Import all libraries needed for the tutorial
import pandas as pd
import pymysql
from pandas.io import sql
from pandas.io.sql import read_sql, to_sql
from sqlalchemy import create_engine, Table, Column, Integer, String, MetaData, ForeignKey

#pd.set_option('display.mpl_style', 'default') 
pd.set_option('display.width', 5000) 
pd.set_option('display.max_columns', 60)
pd.set_option('display.max_rows', 60)

print('is loaded')

base = pd.read_csv('rc15csv/pai2015done.csv', index_col=None)
base.head(3)

base = base.drop('me2015schoolELEM', 1)
base = base.drop('me2015schoolHS', 1)
base = base.drop('zelemscore2015me', 1)
base = base.drop('avgelem2015me', 1)
base = base.drop('stdvelem2015me', 1)
base = base.drop('zelemscore2015meapp', 1)
base = base.drop('avgelem2015meapp', 1)
base = base.drop('stdvelem2015meapp', 1)
base = base.drop('zhsscore2015me', 1)
base = base.drop('avghs2015me', 1)
base = base.drop('stdvhs2015me', 1)
base = base.drop('zhsscore2015meapp', 1)
base = base.drop('avghs2015meapp', 1)
base = base.drop('stdvhs2015meapp', 1)

base = base.applymap(lambda x: "--" if isinstance(x, str) and x.isspace() else x)
base = base.fillna('--')
base.head(5)

base.shape

#with sqlalchemy, create the engine
con = create_engine('mysql+pymysql://root:root@localhost:8889/School Report Cards', echo=False)
print('connected')

metadata = MetaData()
rc2015sch = Table('rc2015sch', metadata,
    Column('schid14', String(20)),
    Column('schid15', String(20)),
    Column('schid', String(20), primary_key=True),
    Column('value', String(100)),
    Column('facilityname', String(85)),
    Column('selective', String(2)),
    Column('districtname', String(40)),
    Column('distid', String(15)),
    Column('distidalt', String(15)),
    Column('county', String(20)),
    Column('districttype', String(15)),
    Column('schooltype', String(21)),
    Column('gradesserved', String(20)),
    Column('administrator', String(35)),
    Column('address', String(35)),
    Column('city', String(35)),
    Column('state', String(2)),
    Column('zip', String(10)),
    Column('telephone', String(12)),
    Column('latitude', String(20)),
    Column('longitude', String(20)),
    Column('note', String(225)),
    Column('me2006schoolisat', String(5)),
    Column('me2006districtisat', String(5)),
    Column('me2006stateisat', String(5)),
    Column('me2007schoolisat', String(5)),
    Column('me2007districtisat', String(5)),
    Column('me2007stateisat', String(5)),
    Column('me2008schoolisat', String(5)),
    Column('me2008districtisat', String(5)),
    Column('me2008stateisat', String(5)),
    Column('me2009schoolisat', String(5)),
    Column('me2009districtisat', String(5)),
    Column('me2009stateisat', String(5)),
    Column('me2010schoolisat', String(5)),
    Column('me2010districtisat', String(5)),
    Column('me2010stateisat', String(5)),
    Column('me2011schoolisat', String(5)),
    Column('me2011districtisat', String(5)),
    Column('me2011stateisat', String(5)),
    Column('me2012schoolisat', String(5)),
    Column('me2012districtisat', String(5)),
    Column('me2012stateisat', String(5)),
    Column('me2013schoolisat', String(5)),
    Column('me2013districtisat', String(5)),
    Column('me2013stateisat', String(5)),
    Column('me2014schoolisat', String(5)),
    Column('me2014districtisat', String(5)),
    Column('me2014stateisat', String(5)),
    Column('me2006schoolpsae', String(5)),
    Column('me2006districtpsae', String(5)),
    Column('me2006statepsae', String(5)),
    Column('me2007schoolpsae', String(5)),
    Column('me2007districtpsae', String(5)),
    Column('me2007statepsae', String(5)),
    Column('me2008schoolpsae', String(5)),
    Column('me2008districtpsae', String(5)),
    Column('me2008statepsae', String(5)),
    Column('me2009schoolpsae', String(5)),
    Column('me2009districtpsae', String(5)),
    Column('me2009statepsae', String(5)),
    Column('me2010schoolpsae', String(5)),
    Column('me2010districtpsae', String(5)),
    Column('me2010statepsae', String(5)),
    Column('me2011schoolpsae', String(5)),
    Column('me2011districtpsae', String(5)),
    Column('me2011statepsae', String(5)),
    Column('me2012schoolpsae', String(5)),
    Column('me2012districtpsae', String(5)),
    Column('me2012statepsae', String(5)),
    Column('me2013schoolpsae', String(5)),
    Column('me2013districtpsae', String(5)),
    Column('me2013statepsae', String(5)),
    Column('me2014schoolpsae', String(5)),
    Column('me2014districtpsae', String(5)),
    Column('me2014statepsae', String(5)),
    Column('act2006school', String(4)),
    Column('act2006district', String(4)),
    Column('act2006state', String(4)),
    Column('act2007school', String(4)),
    Column('act2007district', String(4)),
    Column('act2007state', String(4)),
    Column('act2008school', String(4)),
    Column('act2008district', String(4)),
    Column('act2008state', String(4)),
    Column('act2009school', String(4)),
    Column('act2009district', String(4)),
    Column('act2009state', String(4)),
    Column('act2010school', String(4)),
    Column('act2010district', String(4)),
    Column('act2010state', String(4)),
    Column('act2011school', String(4)),
    Column('act2011district', String(4)),
    Column('act2011state', String(4)),
    Column('act2012school', String(4)),
    Column('act2012district', String(4)),
    Column('act2012state', String(4)),
    Column('act2013school', String(4)),
    Column('act2013district', String(4)),
    Column('act2013state', String(4)),
    Column('act2014school', String(4)),
    Column('actscore2014district', String(4)),
    Column('actscore2014state', String(4)),
    Column('grow2013schoolread', String(5)),
    Column('grow2013distread', String(5)),
    Column('grow2013stateread', String(5)),
    Column('grow2013schoolmath', String(5)),
    Column('grow2013distmath', String(5)),
    Column('grow2013statemath', String(5)),
    Column('grow2014schoolread', String(5)),
    Column('grow2014distread', String(5)),
    Column('grow2014stateread', String(5)),
    Column('grow2014schoolmath', String(5)),
    Column('grow2014distmath', String(5)),
    Column('grow2014statemath', String(5)),
    Column('lowinc2006', String(5)),
    Column('percmin2006', String(5)),
    Column('enroll2006', String(5)),
    Column('enrolldist2006', String(6)),
    Column('grades2006', String(33)),
    Column('zelemscore2006', String(5)),
    Column('avgelem2006', String(5)),
    Column('stdvelem2006', String(5)),
    Column('zhsscore2006', String(5)),
    Column('avghs2006', String(5)),
    Column('stdvhs2006', String(5)),
    Column('lowinc2007', String(5)),
    Column('percmin2007', String(5)),
    Column('enroll2007', String(5)),
    Column('enrolldist2007', String(6)),
    Column('grades2007', String(33)),
    Column('zelemscore2007', String(5)),
    Column('avgelem2007', String(5)),
    Column('stdvelem2007', String(5)),
    Column('zhsscore2007', String(5)),
    Column('avghs2007', String(5)),
    Column('stdvhs2007', String(5)),
    Column('lowinc2008', String(5)),
    Column('percmin2008', String(5)),
    Column('enroll2008', String(5)),
    Column('enrolldist2008', String(6)),
    Column('grades2008', String(33)),
    Column('zelemscore2008', String(5)),
    Column('avgelem2008', String(5)),
    Column('stdvelem2008', String(5)),
    Column('zhsscore2008', String(5)),
    Column('avghs2008', String(5)),
    Column('stdvhs2008', String(5)),
    Column('lowinc2009', String(5)),
    Column('percmin2009', String(5)),
    Column('enroll2009', String(5)),
    Column('enrolldist2009', String(6)),
    Column('grades2009', String(33)),
    Column('zelemscore2009', String(5)),
    Column('avgelem2009', String(5)),
    Column('stdvelem2009', String(5)),
    Column('zhsscore2009', String(5)),
    Column('avghs2009', String(5)),
    Column('stdvhs2009', String(5)),
    Column('lowinc2010', String(5)),
    Column('percmin2010', String(5)),
    Column('enroll2010', String(5)),
    Column('enrolldist2010', String(6)),
    Column('grades2010', String(33)),
    Column('zelemscore2010', String(5)),
    Column('avgelem2010', String(5)),
    Column('stdvelem2010', String(5)),
    Column('zhsscore2010', String(5)),
    Column('avghs2010', String(5)),
    Column('stdvhs2010', String(5)),
    Column('lowinc2011', String(5)),
    Column('percmin2011', String(5)),
    Column('enroll2011', String(5)),
    Column('enrolldist2011', String(6)),
    Column('grades2011', String(33)),
    Column('zelemscore2011', String(5)),
    Column('avgelem2011', String(5)),
    Column('stdvelem2011', String(5)),
    Column('zhsscore2011', String(5)),
    Column('avghs2011', String(5)),
    Column('stdvhs2011', String(5)),
    Column('lowinc2012', String(5)),
    Column('percmin2012', String(5)),
    Column('enroll2012', String(5)),
    Column('enrolldist2012', String(6)),
    Column('grades2012', String(33)),
    Column('zelemscore2012', String(5)),
    Column('avgelem2012', String(5)),
    Column('stdvelem2012', String(5)),
    Column('zhsscore2012', String(5)),
    Column('avghs2012', String(5)),
    Column('stdvhs2012', String(5)),
    Column('lowinc2013', String(5)),
    Column('percmin2013', String(5)),
    Column('enroll2013', String(5)),
    Column('enrolldist2013', String(6)),
    Column('grades2013', String(33)),
    Column('zelemscore2013', String(5)),
    Column('avgelem2013', String(5)),
    Column('stdvelem2013', String(5)),
    Column('zhsscore2013', String(5)),
    Column('avghs2013', String(5)),
    Column('stdvhs2013', String(5)),
    Column('lowinc2014', String(5)),
    Column('percmin2014', String(5)),
    Column('enroll2014', String(5)),
    Column('enrolldist2014', String(6)),
    Column('grades2014', String(33)),
    Column('zelemscore2014', String(5)),
    Column('avgelem2014', String(5)),
    Column('stdvelem2014', String(5)),
    Column('zhsscore2014', String(5)),
    Column('avghs2014', String(5)),
    Column('stdvhs2014', String(5)),
    Column('grades2015', String(33)),
    Column('lowinc2015', String(5)),
    Column('percmin2015', String(5)),
    Column('enroll2015', String(7)),
    Column('enrolldist2015', String(8)),
    Column('act2015school', String(4)),
    Column('actscore2015district', String(4)),
    Column('actscore2015state', String(4)),
    Column('me2015school', String(5)),
    Column('me2015district', String(5)),
    Column('me2015state', String(5)),
    Column('me2015schDNY', String(5)),
    Column('me2015schPART', String(5)),
    Column('me2015schAPP', String(5)),
    Column('me2015schMET', String(5)),
    Column('me2015schEXC', String(5)),
    Column('me2015distDNY', String(5)),
    Column('me2015distPART', String(5)),
    Column('me2015distAPP', String(5)),
    Column('me2015distMET', String(5)),
    Column('me2015distEXC', String(5)),
    Column('me2015stateDNY', String(5)),
    Column('me2015statePART', String(5)),
    Column('me2015stateAPP', String(5)),
    Column('me2015stateMET', String(5)),
    Column('me2015stateEXC', String(5)),
    Column('avgELEMz', String(5)),
    Column('avgHSz', String(5)),
    mysql_engine='InnoDB')

con.execute("DROP TABLE IF EXISTS rc2015sch")
metadata.create_all(con)
print("created")

base.to_sql(name='rc2015sch', con=con, if_exists = 'append', index=False, index_label=None, chunksize=1000)
print('done')
